requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
pyarrow>=14.0.0
plotly>=5.18.0
scikit-learn>=1.4.0
xgboost>=2.0.0
joblib>=1.3.0
google-generativeai>=0.5.0
requests>=2.31.0
python-dotenv>=1.0.0
